Struct hashmap2::VacantEntry

pub struct VacantEntry<'a, K: 'a, V: 'a> {
    // some fields omitted
}

A view into a single empty location in a HashMap.

Methods

impl<'a, K: 'a, V: 'a> VacantEntry<'a, K, V>

fn insert(self, value: V) -> &'a mut V

Sets the value of the entry with the VacantEntry's key, and returns a mutable reference to it

fn key(&self) -> &K

Gets a reference to the entry key

Examples

use hashmap2::HashMap;

let mut map = HashMap::<String, u32>::new();

assert_eq!("foo", map.entry("foo".to_string()).key());